Moving page to new location to allow for better journey
