+++ app/navigation/PreTaskListNavigator.scala
@@ -24,20 +24,15 @@ import play.api.mvc.Call
 
 @Singleton
 class PreTaskListNavigator @Inject()() extends Navigator {
-
+  // format: off
   override val normalRoutes: PartialFunction[Page, UserAnswers => Option[Call]] = {
-    case LocalReferenceNumberPage =>
-      ua =>
-        Some(routes.AddSecurityDetailsController.onPageLoad(ua.id, NormalMode))
-    case AddSecurityDetailsPage =>
-      ua =>
-        Some(routes.DeclarationSummaryController.onPageLoad(ua.id))
+    case LocalReferenceNumberPage => ua => Some(routes.ProcedureTypeController.onPageLoad(ua.id, NormalMode))
+    case ProcedureTypePage => ua => Some(routes.AddSecurityDetailsController.onPageLoad(ua.id, NormalMode))
+    case AddSecurityDetailsPage => ua => Some(routes.DeclarationSummaryController.onPageLoad(ua.id))
   }
 
   override val checkRoutes: PartialFunction[Page, UserAnswers => Option[Call]] = {
-    case _ =>
-      _ =>
-        None
+    case _ => _ => None
   }
-
+  // format: on
 }
