+++ app/utils/MovementDetailsCheckYourAnswersHelper.scala
@@ -104,22 +104,6 @@ class MovementDetailsCheckYourAnswersHelper(userAnswers: UserAnswers) {
       )
   }
 
-  def procedureType: Option[Row] = userAnswers.get(ProcedureTypePage) map {
-    answer =>
-      Row(
-        key   = Key(msg"procedureType.checkYourAnswersLabel", classes = Seq("govuk-!-width-one-half")),
-        value = Value(msg"procedureType.$answer"),
-        actions = List(
-          Action(
-            content            = msg"site.edit",
-            href               = routes.ProcedureTypeController.onPageLoad(lrn, CheckMode).url,
-            visuallyHiddenText = Some(msg"site.edit.hidden".withArgs(msg"procedureType.checkYourAnswersLabel")),
-            attributes         = Map("id" -> s"""change-procedure-type""")
-          )
-        )
-      )
-  }
-
   def declarationType: Option[Row] = userAnswers.get(DeclarationTypePage) map {
     answer =>
       Row(
