+++ conf/views/procedureType.njk
@@ -27,7 +27,7 @@
           }) }}
         {% endif %}
 
-        <form action="{{ routes.controllers.movementDetails.ProcedureTypeController.onSubmit(lrn, mode).url }}" method="post">
+        <form action="{{ routes.controllers.ProcedureTypeController.onSubmit(lrn, mode).url }}" method="post">
 
         <span class="govuk-caption-xl">{{messages('movementDetails.preHeading')}}</span>
